move auth and tenant provider into new api module
